// lib/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
      items: {
        Row: {
          created_at: string
          have_bought: boolean
          id: string
          list_id: string
          name: string
        }
        Insert: {
          created_at?: string
          have_bought?: boolean
          id?: string
          list_id: string
          name: string
        }
        Update: {
          created_at?: string
          have_bought?: boolean
          id?: string
          list_id?: string
          name?: string
        }
        Relationships: [
          {
            foreignKeyName: "items_list_id_fkey"
            columns: ["list_id"]
            referencedRelation: "lists"
            referencedColumns: ["id"]
          }
        ]
      }
      lists: {
        Row: {
          created_at: string
          id: string
          title: string
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          title: string
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          title?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "lists_user_id_fkey"
            columns: ["user_id"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}